StaticAndFinal
Here a dice game is simulated with three players where the result of a single die cast is randomly generated. The static keyword designates that variable to be constant, meaning there is only one instance at any time hence the variable diceRoll & diceRollCount can safely be made static. the final keyword designates the data value of a variable to be unchangeable. Since the score to win is 60, it has been set to static and final.

A few get methods were created for the purpose of getting the roll & score value to be output after each player casts a die, to let the user see what is happening.

Ultimately an infinitely looping while statement is used with three nested if statements, one for each player, that terminates the loop once any player score reaches or exceeds 60. The first player to do so wins the game.

Also embedded into the Roll method is a counter which counts the number of times the Roll method was called (how many times the die was cast) in total, and lets the user know this information upon game win.